Moved odc.index.stac => odc.stac.transform
