+++ apps/dc_tools/odc/apps/dc_tools/fs_to_dc.py
@@ -10,7 +10,7 @@ from odc.apps.dc_tools.utils import (
     allow_unsafe,
     transform_stac,
 )
-from odc.index.stac import stac_transform
+from odc.stac.transform import stac_transform
 from typing import Generator, Optional
 import logging
