+++ apps/dc_tools/odc/apps/dc_tools/index_from_tar.py
@@ -7,7 +7,7 @@ import click
 import datacube
 from datacube.utils.changes import allow_any
 from odc.index import from_yaml_doc_stream
-from odc.index.stac import stac_transform
+from odc.stac.transform import stac_transform
 from odc.io.tar import tar_doc_stream, tar_mode
 from odc.io.timer import RateEstimator
