+++ apps/dc_tools/odc/apps/dc_tools/stac_api_to_dc.py
@@ -16,7 +16,7 @@ from odc.apps.dc_tools.utils import (
     limit,
     update_if_exists,
 )
-from odc.index.stac import stac_transform, stac_transform_absolute
+from odc.stac.transform import stac_transform, stac_transform_absolute
 from pystac.item import Item
 from pystac_client import Client
